client

+++ app/core/views/clients.py
@@ -13,11 +13,12 @@ __all__ = ["search_client", "client_list", "client_add", "delete_clients"]
 
 def delete_clients(request):
     ids = request.POST.getlist("ids")
-    Client.objects.filter(id__in=ids).delete()
-    resp = HttpResponse("")
-    messages.info(request, _("Clients deleted"))
-    resp["X-IC-Redirect"] = reverse("client_list")
-    return resp
+    if ids:
+        Client.objects.filter(id__in=ids).delete()
+        resp = HttpResponse("")
+        messages.info(request, _("Clients deleted"))
+        resp["X-IC-Redirect"] = reverse("client_list")
+        return resp
